requirements.txt
kiteconnect==5.0.1
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
python-dotenv==1.0.1
requests==2.32.3
